qt_gl_gst: Updates to get demo running on OMAP3530EVM with OpenGL|ES. Re-arranged shaders and loader to concatenate before compiling, as SGX doesn't link multiple shaders of the same type. Added UYVY format conversion fragment shaders.

+++ src/qt_gl_gst/shaders/noeffect.vert
@@ -8,7 +8,7 @@ uniform highp mat4 u_mv_matrix;
 attribute highp vec4 a_vertex;
 attribute highp vec4 a_texCoord;
 
-varying highp vec4 v_texCoord;
+varying mediump vec4 v_texCoord;
 
 void main(void)
 {
